Change the default behaviour of all scripts to NOT fetch.
This already was the default behaviour of git-flow-feature, but now it
is the default for the other scripts, too.

RATIONALE: Due to limitations on some platforms (and some
implementations of getopt), it's impossible to turn off the -f (fetch)
option.  Therefore, it must now be set explicitly.

Also, this makes git-flow work in stand-alone repositories (i.e. repos
that do not have an origin remote at all).