Don't assume the number of supporting branches is static.
Branch managers are autodetected already, so do the same with the
prefixes.  Instead of hardcoded feature_prefix() methods, there now is
a general get_prefix() method that reads the prefix for the given branch
manager's identifier (i.e. 'feature', 'release', 'hotfix', etc.)

This makes it easy to extend the git-flow model with more types of
branches in the future, simply by implementing extra subclasses of
BranchManager.