Escape characters from tags that have special meaning in grep.
In particular, this fixes the case where a dot in a version name (not
too uncommon ;)) unintentionally matches a pre-existing tag.  For
example, if these tags exist:

    1.0.1
    1.0b2

And you try to start a new release for 1.0.2, git-flow prevented it,
since the '.' matches the 'b' in 1.0b2.  This resulted in the invalid
error message:

    Tag '1.0.2' already exists. Pick another name.