Change tag sort and auto-annotate created tags.
Tag sort is now done by date on the tag, not alpha. This should
ensure we're always looking at the latest tag when incrementing
version number while remaining much simpler than a version sort.

Tags that are created by feature finish and hotfix finish are now
auto-annotated with the version number. This can still be
overridden by adding a -m flag with message to the finish call,
but avoids an unnecessary interruption since the actual tag
message is mostly worthless, but we have to have it to get
author and date metadata.