Support setting "aliases", so one can have multiple workflows in the same tree
To support multiple workflows in the same tree, allow the use of
either symlinks or GIT_FLOW_SELF to select between flows.

This allows one to, for example, set GIT_FLOW_SELF to git-flow-debian,
in which case git flow will look for the gitflow-debian settings in
git config, instead of gitflow.

This, in turn, makes it possible to have a different workflow setup
for upstream development, and packaging development.

All this is accomplished with the introduction of a gitflow_config
function, that does the heavy lifting, and everything that used git
config gitflow.*, will now use git config $(gitflow_config).*.

Requested-By: Holger Levsen <holger@layer-acht.org>
Signed-off-by: Gergely Nagy <algernon@balabit.hu>