Fixes the suggested name for the integration branch.
When git init -d is executed on a repo with multiple branches, the suggested
name for the integration branch could possibly be the same name that was
already selected for the production release branch.

If no suggested name is determined the suggested name remains empty which
leads to all sorts of other problems.